requirements.txt
anyio==3.3.4
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
ecdsa==0.17.0
fastapi==0.70.0
h11==0.12.0
httptools==0.2.0
idna==3.3
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.0
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
motor==2.2.0
requests
python-multipart